service: use 'additional dns hostnames' with net ads join
With newer versions of Samba the net ads join does not add services
principals with the configured host name anymore but added the new
option 'additional dns hostnames' for this.

realmd will try to figure out a fully-qualified host name and use it
with the new option if it is from a different domain.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1867912