paths: install realmd in libexecdir
The realmd binary will be installed in autoconf's libexecdir. Since it
is the only binary installed in libexecdir the location of realmd can be
selected with the --libexecdir confiure option.

A new configure option --with-private-dir was added to allow to specify
to directory for the default configuration files. The default is
$prefix/lib/realmd since the distro defaults might contain arch specific
scripts.

Resolves: https://gitlab.freedesktop.org/realmd/realmd/-/issues/26